refactor: Separate set saved case into seperate hook.
